Refactor 'ResourceKey' for the measurement public api.
