+++ src/main/kotlin/org/wfanet/measurement/api/v2alpha/MeasurementConsumerKey.kt
@@ -15,6 +15,7 @@
 package org.wfanet.measurement.api.v2alpha
 
 import org.wfanet.measurement.common.ResourceNameParser
+import org.wfanet.measurement.common.api.ResourceKey
 
 /** [ResourceKey] of a Measurement Consumer. */
 data class MeasurementConsumerKey(val measurementConsumerId: String) : ResourceKey {
