+++ src/main/kotlin/org/wfanet/measurement/api/v2alpha/ModelProviderKey.kt
@@ -15,6 +15,7 @@
 package org.wfanet.measurement.api.v2alpha
 
 import org.wfanet.measurement.common.ResourceNameParser
+import org.wfanet.measurement.common.api.ResourceKey
 
 /** [ModelProviderKey] of a Model Provider. */
 data class ModelProviderKey(val modelProviderId: String) : ResourceKey {
