+++ src/main/kotlin/org/wfanet/measurement/api/v2alpha/RecurringExchangeKey.kt
@@ -15,6 +15,7 @@
 package org.wfanet.measurement.api.v2alpha
 
 import org.wfanet.measurement.common.ResourceNameParser
+import org.wfanet.measurement.common.api.ResourceKey
 
 private val parser = ResourceNameParser("recurringExchanges/{recurring_exchange}")
