+++ src/main/kotlin/org/wfanet/measurement/common/api/ResourceKey.kt
@@ -1,4 +1,4 @@
-// Copyright 2021 The Cross-Media Measurement Authors
+// Copyright 2022 The Cross-Media Measurement Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.measurement.api.v2alpha
+package org.wfanet.measurement.common.api
 
 interface ResourceKey {
   /** Converts this [ResourceKey] into a resource name. */
